pyproject.toml
[tool.poetry]
name = "gpt-custom-chatbot"
version = "0.1.0"
description = ""
authors = ["Nguyen Cuong <36143842+ntcuong777@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
streamlit = "^1.37.1"
pyyaml = "^6.0.2"
cutelog = "^2.2.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.32"
openai = "^1.42.0"
scipy = "^1.14.1"
tiktoken = "^0.7.0"
anthropic = "^0.34.1"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
numpy = "<2.0.0"
pydantic-settings = "^2.4.0"
snowflake-id = "^1.0.2"
streamlit-markdown = "^1.0.4"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
